modify: sticky 네비게이션 바 메인페이지에만 적용
